LPS-104241 SF, move classes. Non-base classes implementing StrutsAction needs to be in package ending with "struts"
